requirements.txt
# requirements.txt - Excel Edition
# Bot Telegram
python-telegram-bot==20.7

# Dashboard Streamlit
streamlit==1.29.0
plotly==5.17.0

# Excel/Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# Configurações e Logging
PyYAML==6.0.1
python-dotenv==1.0.0

# Utilitários
python-dateutil==2.8.2
tzlocal==5.2